Use T_START/T_END marker lumps

+++ r_local.h
@@ -388,7 +388,7 @@ extern int8_t* dc_colormaps;
 extern uint8_t* dc_playpals;
 
 #ifdef MARS
-#define R_CheckPixels(lumpnum) (void *)((intptr_t)(W_POINTLUMPNUM(lumpnum)))
+#define R_CheckPixels(lumpnum) (void *)(W_POINTLUMPNUM(lumpnum))
 #endif
 
 void R_InitTextures(void);
